requirements.txt
Django==5.2.5
djangorestframework==3.16.0
drf-yasg==1.21.10
PyJWT==2.10.1
drf-mocker==0.1.0
